requirements.txt
python-telegram-bot>=21.0
telethon==1.34.0
openai>=1.12.0
requests>=2.31.0
python-dotenv==1.0.0
phonenumbers>=8.13.0

